Refactor code

+++ RidesService/src/main/java/by/modsen/taxiprovider/ridesservice/dto/ride/PotentialCostDTO.java
@@ -4,10 +4,12 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class PotentialCostDTO {
 
-    private double cost;
+    private BigDecimal cost;
 }
